Began major redesign work
The idea here is to have separate book formatting modules for everything, more uniformly handle authors and titles that belong together and to factor out more reusable code into common modules. Another goal of this is to provide an easier framework to add support for new book formatters without upheaving existing formatters.
Another part of this effort is to perform unit testing as an integral part of redesign. When we're done we'll have a comprehensive set of unit tests in place.