Added back commented-out ghc-options for Windows binary strip
Turns out binaries are not automatically stripped on Windows with Haskell Platform 2011.2.0.1 regardless of what cabal sdist and Hackage say about it. Trying to use if os(windows) blocks in the cabal file gets rejected by Hackage. Workaround is to make the developer uncomment the lines with -optl-s if building for Windows.