OPF file import code now using strict bytestrings
Changed to using strict bytestrings for
codec.Epub.IO.opfContentsFromBS call, and had to make corresponding
type/code changes here.

This entire set of changes (in this project and epub-metadata) arose
because of a bug found under Windows. When renaming with epubname,
the epub-metadata library use of lazy bytestrings was keeping the
epub file open, preventing rename. All of this has been changed to
be strict on the epub file read.