Big redesign of EN monad
Was able to do a couple of big things in one large redesign. First, removed a lot of boilerplate passing around of book Metadata by putting it together with the Options Reader state in the EN monad. This meant that it was necessary to perform the book file metadata load earlier. And that led to factoring all IO entirely out of the EN monad. This is better design.