Major redesign of formatting rules system
This software has undergone a very large redesign. The rules and
machinery for recognizing a specific publication and determining the
proper new filename are no longer statically compiled in Haskell source
code. A domain-specific langauge has been invented for this, intended
for end-users of this program. They are now able to add new filename
formatters for anything they need.