Large redesign of epubname main code
Two big things were done here:

First, a lot of code to initialize the name formatting process and do the formatting has been pushed out of epubname.hs so it's more sensible and smaller to re-use (as in epubzip) This involved a lot of rethinking of what code is responsible for what.

Second, a lot of ugly, nested return value checking code in epubname.hs was avoided by redesigning many things to work within ErrorT monad.