Removed ordinaryBookFormatter static definition
All software that needs to access the book renaming machinery now must
load a rules file from disk at runtime. This required some big changes to
the epubzip utility in this project. But I think it's worth it to be able
to auto-generate book names for a lot more books with that tool as well.