Removed --overwrite option
Turns out, something (the filesystem? the System.Directory module?) is
smart enough to not do a thing when you renameFile with the same name
it already has. This switch has been pointless all along.

Also changed the wording of the file exists output to seem less like a
big problem.