Fixed a rule that was failing on Windows
I believe this was two things. First an icorrectly escaped backslash-s in
the expression that was somehow working properly in Linux. Second, once I
did esape the character properly, this didn't work in Windows. Thinking
that regex-compat on that platform isn't compatible with the whitespace
notation. Fix was changing it to one-or-more literal spaces: " +"