Fixed stack overflow problem with reading some epub files
Found that some epub documents were causing a stack overflow in
the zip-archive library we use. Was able to fix this by changing
the stack size from the default 8M to 32M. I believe this to be a
generous change, as 12M was enough to fix the problem with the books
I have access to. However, to be safe, the relevant binaries also
have the -rtsopts switch on, so adjustments can be made by users
as necessary.