Restored --any-date switch and behavior specific to epub version
Turns out that we need to handle publication date differently for
epub2 and epub3 documents. So I ended up putting the --any-date
switch back. This required some deeper redesign to stuff the epub
Package data structure into the Globals data type that becomes the
state for the EN monad.

Extensive unit test editing was required, and the usage text was
updated to reflect the new behavior.