Fixed an error in bad DSL command index reporting
We have to subtract 1 from the indices in the rules file because
the lists of matches are 0-based (as is customary in code), but I
wanted to treat these matches as 1-based as is customary in Perl
and most regexp libraries. So when there was an invalid index used,
the number reported back was one less than the number in the rules
file. Fixed this.