Fixed a bug in epub opf editing
The code in EpubTools.EpubMeta.Import was writing the new OPF file into
the new book as the first zip file entry. This is invalid, the mimetype
file must be the first zip entry in an epub book.

This has been fixed by writing both files into the new archive in the
correct order.