Fixed special "info" type cli arguments
--dump-rules and --help-rules were not working with the new
change to optparse-applicative because we require at least one
non-switched argument for the options to successfully parse.

The solution to this was to take those two options out of the Options
data structure entirely and do the same as the --version switch. These
infoOption type parsers are now picked off by execParser in
optparse-applicative and never get to main at all.