Completely overhauled options parsing for epubmeta
This switches over from GetOpt to optparse-applicative

We now use a mode-like set of switches like pacman or apt with many
mode-specific options in each mode.

Part of this change was also adding more places backup copies of the
epub file can be specified by the user. A specific mode for viewing the
metadata. And a lot more types everywhere.